1070 report tests (#349)
* altered to not use property as lint raised it as a static token


* adding testing snapshots to the reports service.
