+++ legal-api/src/legal_api/reports/__init__.py
@@ -9,8 +9,10 @@
 # an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 """Module to manage the calls and content to the reporting service."""
-
 from http import HTTPStatus
+
+from flask import jsonify
+from flask_babel import _
 
 from .report import Report
 
@@ -21,4 +23,4 @@ def get_pdf(filing):
         return Report(filing).get_pdf()
     except FileNotFoundError:
         # We don't have a template for it, so it must only be available on paper.
-        return 'Available on paper only', HTTPStatus.NOT_FOUND
+        return jsonify({'message': _('Available on paper only.')}), HTTPStatus.NOT_FOUND
